refactor: move stores into getDocumentTitle fn
This makes this function and it's use more simple since you don't need
to get the stores before calling it.

This change also changes the name of the function and file name to
reflect the usage better.
